cmd: break out bucket cmds into files
