+++ cmd/buck/cli/links.go
@@ -1,0 +1,32 @@
+package cli
+
+import (
+	"github.com/spf13/cobra"
+	"github.com/textileio/textile/cmd"
+)
+
+var bucketLinksCmd = &cobra.Command{
+	Use: "links",
+	Aliases: []string{
+		"link",
+	},
+	Short: "Show links to where this bucket can be accessed",
+	Long:  `Displays a thread, IPNS, and website link to this bucket.`,
+	Args:  cobra.ExactArgs(0),
+	PreRun: func(c *cobra.Command, args []string) {
+		cmd.ExpandConfigVars(config.Viper, config.Flags)
+		if config.Viper.ConfigFileUsed() == "" {
+			cmd.Fatal(errNotABucket)
+		}
+	},
+	Run: func(c *cobra.Command, args []string) {
+		ctx, cancel := clients.Ctx.Thread(cmd.Timeout)
+		defer cancel()
+		key := config.Viper.GetString("key")
+		reply, err := clients.Buckets.Links(ctx, key)
+		if err != nil {
+			cmd.Fatal(err)
+		}
+		printLinks(reply)
+	},
+}
