+++ cmd/buck/cli/status.go
@@ -1,0 +1,42 @@
+package cli
+
+import (
+	"path/filepath"
+
+	"github.com/ipfs/interface-go-ipfs-core/options"
+	"github.com/spf13/cobra"
+	"github.com/textileio/textile/buckets/local"
+	"github.com/textileio/textile/cmd"
+)
+
+var bucketStatusCmd = &cobra.Command{
+	Use: "status",
+	Aliases: []string{
+		"st",
+	},
+	Short: "Show bucket object changes",
+	Long:  `Displays paths that have been added to and paths that have been removed or differ from the local bucket root.`,
+	Args:  cobra.ExactArgs(0),
+	PreRun: func(c *cobra.Command, args []string) {
+		cmd.ExpandConfigVars(config.Viper, config.Flags)
+	},
+	Run: func(c *cobra.Command, args []string) {
+		conf := config.Viper.ConfigFileUsed()
+		if conf == "" {
+			cmd.Fatal(errNotABucket)
+		}
+		root := filepath.Dir(filepath.Dir(conf))
+		buck, err := local.NewBucket(root, options.BalancedLayout)
+		if err != nil {
+			cmd.Fatal(err)
+		}
+		diff := getDiff(buck, root)
+		if len(diff) == 0 {
+			cmd.End("Everything up-to-date")
+		}
+		for _, c := range diff {
+			cf := changeColor(c.Type)
+			cmd.Message("%s  %s", cf(changeType(c.Type)), cf(c.Rel))
+		}
+	},
+}
